+++ tests/integration/test_custom_application_source_version.py
@@ -18,10 +18,11 @@ from typing import Any, Generator
 import pytest
 
 import datarobot as dr
-from datarobot.models.runtime_parameters import RuntimeParameterValue
 from datarobotx.idp.credentials import get_replace_or_create_credential  # type: ignore
 from datarobotx.idp.custom_application_source import (  # type: ignore
     get_or_create_custom_application_source,
+)
+from datarobotx.idp.custom_application_source_version import (  # type: ignore
     get_or_create_custom_application_source_version,
     get_or_create_custom_application_source_version_from_previous,
 )
@@ -87,7 +88,7 @@ def credentials(dr_endpoint: str, dr_token: str, cleanup_dr: Any) -> Generator[s
 
 @pytest.fixture
 def app_context_path(
-    tmp_path: pathlib.Path, requirements: str, app_py: str, start_script: str
+    tmp_path: pathlib.Path, requirements: str, app_py: str, start_script: str, metadata: str
 ) -> str:
     ts = 1701797283
     folder_path = tmp_path / "app"
@@ -101,16 +102,21 @@ def app_context_path(
     p = folder_path / "start-app.sh"
     p.write_text(start_script)
     os.utime(p, (ts, ts))
+    p = folder_path / "metadata.yaml"
+    p.write_text(metadata)
+    os.utime(p, (ts, ts))
     return str(folder_path.resolve())
 
 
 @pytest.fixture
-def additional_metadata_context_path(tmp_path: pathlib.Path, metadata: str) -> str:
+def additional_metadata_context_path(
+    tmp_path: pathlib.Path,
+) -> str:
     ts = 1701797283
     folder_path = tmp_path / "additional_app"
     folder_path.mkdir()
-    p = folder_path / "metadata.yaml"
-    p.write_text(metadata)
+    p = folder_path / "foo.txt"
+    p.write_text("foo")
     os.utime(p, (ts, ts))
     return str(folder_path.resolve())
 
@@ -124,6 +130,18 @@ def custom_application_source(
             endpoint=dr_endpoint,
             token=dr_token,
             name="pytest custom application source",
+        )
+
+
+@pytest.fixture
+def custom_application_source_2(
+    dr_endpoint: str, dr_token: str, cleanup_dr: Any
+) -> Generator[str, None, None]:
+    with cleanup_dr("customApplicationSources/"):
+        yield get_or_create_custom_application_source(
+            endpoint=dr_endpoint,
+            token=dr_token,
+            name="pytest custom application source_2",
         )
 
 
@@ -142,7 +160,6 @@ def test_get_or_create(
     additional_metadata_context_path: str,
     credentials: str,
 ) -> None:
-    client = dr.Client(endpoint=dr_endpoint, token=dr_token)  # type: ignore
     env_ver_id_1 = get_or_create_custom_application_source_version(
         endpoint=dr_endpoint,
         token=dr_token,
@@ -169,64 +186,85 @@ def test_get_or_create(
         label="pytest app source version alt",
     )
     assert env_ver_id_1 != env_ver_id_3
-    folder_contents = [
-        item["fileName"]
-        for item in client.get(
-            f"customApplicationSources/{custom_application_source}/versions/{env_ver_id_3}"
-        ).json()["items"]
-    ]
-    env_ver_id_4 = get_or_create_custom_application_source_version_from_previous(
+
+
+def test_get_or_create_with_runtime_params(
+    dr_endpoint: str,
+    dr_token: str,
+    cleanup_env_ver: Any,
+    custom_application_source_2: str,
+    app_context_path: str,
+    additional_metadata_context_path: str,
+    credentials: str,
+) -> None:
+    client = dr.Client(endpoint=dr_endpoint, token=dr_token)  # type: ignore
+
+    env_ver_id_1 = get_or_create_custom_application_source_version(
         endpoint=dr_endpoint,
         token=dr_token,
-        custom_application_source_id=custom_application_source,
-        # previous_custom_application_source_version_id=env_ver_id_3,
-        folder_path=additional_metadata_context_path,
-    )
-    assert env_ver_id_3 != env_ver_id_4
-    folder_contents_new = [
-        item["fileName"]
-        for item in client.get(
-            f"customApplicationSources/{custom_application_source}/versions/{env_ver_id_4}"
-        ).json()["items"]
-    ]
-    assert set(folder_contents + ["metadata.yaml"]) == set(folder_contents_new)
-    env_ver_id_5 = get_or_create_custom_application_source_version_from_previous(
-        endpoint=dr_endpoint,
-        token=dr_token,
-        custom_application_source_id=custom_application_source,
-        # previous_custom_application_source_version_id=env_ver_id_4,
+        custom_application_source_id=custom_application_source_2,
         runtime_parameter_values=[
             {"field_name": "SOME_TEXT", "type": "string", "value": "foo"},
             {"field_name": "SOME_CREDENTIALS", "type": "credential", "value": credentials},
         ],
+        folder_path=app_context_path,
     )
-    assert env_ver_id_4 != env_ver_id_5
+
+    assert env_ver_id_1
+
     runtime_params = client.get(
-        f"customApplicationSources/{custom_application_source}/versions/{env_ver_id_5}"
+        f"customApplicationSources/{custom_application_source_2}/versions/{env_ver_id_1}"
     ).json()["runtimeParameters"]
     some_text_metadata = next(item for item in runtime_params if item["fieldName"] == "SOME_TEXT")
-    assert some_text_metadata["currentValue"] == "foo"
     some_creds_metadata = next(
         item for item in runtime_params if item["fieldName"] == "SOME_CREDENTIALS"
     )
+
+    assert some_text_metadata["currentValue"] == "foo"
     assert some_creds_metadata["currentValue"] == credentials
 
-    env_ver_id_6 = get_or_create_custom_application_source_version_from_previous(
+    env_ver_id_2 = get_or_create_custom_application_source_version(
         endpoint=dr_endpoint,
         token=dr_token,
-        custom_application_source_id=custom_application_source,
-        # previous_custom_application_source_version_id=env_ver_id_4,
+        custom_application_source_id=custom_application_source_2,
         runtime_parameter_values=[
-            RuntimeParameterValue(field_name="SOME_TEXT", value="bar", type="string")
+            {"field_name": "SOME_TEXT", "type": "string", "value": "foo"},
+            {"field_name": "SOME_CREDENTIALS", "type": "credential", "value": credentials},
         ],
+        folder_path=app_context_path,
     )
-    assert env_ver_id_4 != env_ver_id_6
-    runtime_params = client.get(
-        f"customApplicationSources/{custom_application_source}/versions/{env_ver_id_6}"
-    ).json()["runtimeParameters"]
-    some_text_metadata = next(item for item in runtime_params if item["fieldName"] == "SOME_TEXT")
-    assert some_text_metadata["currentValue"] == "bar"
-    some_creds_metadata = next(
-        item for item in runtime_params if item["fieldName"] == "SOME_CREDENTIALS"
+
+    assert env_ver_id_1 == env_ver_id_2
+
+    folder_contents_old = [
+        item["fileName"]
+        for item in client.get(
+            f"customApplicationSources/{custom_application_source_2}/versions/{env_ver_id_2}"
+        ).json()["items"]
+    ]
+
+    env_ver_id_3 = get_or_create_custom_application_source_version_from_previous(
+        endpoint=dr_endpoint,
+        token=dr_token,
+        custom_application_source_id=custom_application_source_2,
+        folder_path=additional_metadata_context_path,
     )
-    assert some_creds_metadata["currentValue"] == credentials
+    assert env_ver_id_3 != env_ver_id_2
+
+    folder_contents_new = [
+        item["fileName"]
+        for item in client.get(
+            f"customApplicationSources/{custom_application_source_2}/versions/{env_ver_id_3}"
+        ).json()["items"]
+    ]
+
+    assert folder_contents_old + ["foo.txt"] == folder_contents_new
+
+    env_ver_id_4 = get_or_create_custom_application_source_version_from_previous(
+        endpoint=dr_endpoint,
+        token=dr_token,
+        custom_application_source_id=custom_application_source_2,
+        folder_path=additional_metadata_context_path,
+    )
+
+    assert env_ver_id_3 == env_ver_id_4
