+++ tests/integration/test_custom_applications.py
@@ -17,19 +17,23 @@ from typing import Any, Generator
 
 import pytest
 
-from datarobotx.idp.custom_application_source import (
+from datarobotx.idp.custom_application_source import (  # type: ignore
     get_or_create_custom_application_source,
+)
+from datarobotx.idp.custom_application_source_version import (  # type: ignore
     get_or_create_custom_application_source_version,
 )
-from datarobotx.idp.custom_applications import (
+from datarobotx.idp.custom_applications import (  # type: ignore
     _find_existing_custom_app,
     get_replace_or_create_custom_app,
     get_replace_or_create_custom_app_from_env,
 )
-from datarobotx.idp.execution_environment_versions import (
+from datarobotx.idp.execution_environment_versions import (  # type: ignore
     get_or_create_execution_environment_version,
 )
-from datarobotx.idp.execution_environments import get_or_create_execution_environment
+from datarobotx.idp.execution_environments import (  # type: ignore
+    get_or_create_execution_environment,
+)
 
 
 @pytest.fixture()
